Basic Drive Train Setup
While the code copied over from Team 4069 looks promising, it will still
need tweaking to understand/finish. So, I split the repository into
branches. This branch uses the basic drive train code that we wrote
before everything went south. We will keep this as the master branch
until we are certain that we have PID code that works.